requirements_render.txt
python-telegram-bot==13.15
mercadopago>=2.2.0
python-dotenv>=0.21.0
requests>=2.28.0
flask>=2.2.0
gunicorn>=20.1.0